Create customization arg utils
