+++ core/domain/state_domain.py
@@ -20,12 +20,12 @@ import copy
 import logging
 
 from constants import constants
+from core.domain import customization_args_util
 from core.domain import html_cleaner
 from core.domain import interaction_registry
 from core.domain import param_domain
 import feconf
 import jinja_utils
-import schema_utils
 import utils
 
 
@@ -270,7 +270,7 @@ class InteractionInstance(object):
             'id': self.id,
             'customization_args': (
                 {} if self.id is None else
-                schema_utils.CustomizationArgsUtil.get_full_customization_args(
+                customization_args_util.get_full_customization_args(
                     self.customization_args,
                     interaction_registry.Registry.get_interaction_by_id(
                         self.id).customization_arg_specs)),
@@ -399,7 +399,7 @@ class InteractionInstance(object):
         except KeyError:
             raise utils.ValidationError('Invalid interaction id: %s' % self.id)
 
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
+        customization_args_util.validate_customization_args_and_values(
             'interaction', self.id, self.customization_args,
             interaction.customization_arg_specs)
