+++ core/domain/stats_domain.py
@@ -20,12 +20,12 @@ import numbers
 import sys
 
 from core.domain import action_registry
+from core.domain import customization_args_util
 from core.domain import exp_domain
 from core.domain import interaction_registry
 from core.domain import playthrough_issue_registry
 from core.platform import models
 import feconf
-import schema_utils
 import utils
 
 (stats_models,) = models.Registry.import_models([models.NAMES.statistics])
@@ -615,7 +615,7 @@ class Playthrough(object):
             raise utils.ValidationError('Invalid issue type: %s' % (
                 self.issue_type))
 
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
+        customization_args_util.validate_customization_args_and_values(
             'issue', self.issue_type, self.issue_customization_args,
             issue.customization_arg_specs)
 
@@ -662,7 +662,7 @@ class ExplorationIssue(object):
         return {
             'issue_type': self.issue_type,
             'issue_customization_args': (
-                schema_utils.CustomizationArgsUtil.get_full_customization_args(
+                customization_args_util.get_full_customization_args(
                     self.issue_customization_args,
                     playthrough_issue_registry.Registry.get_issue_by_type(
                         self.issue_type).customization_arg_specs)),
@@ -760,7 +760,7 @@ class ExplorationIssue(object):
             raise utils.ValidationError('Invalid issue type: %s' % (
                 self.issue_type))
 
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
+        customization_args_util.validate_customization_args_and_values(
             'issue', self.issue_type, self.issue_customization_args,
             issue.customization_arg_specs)
 
@@ -803,7 +803,7 @@ class LearnerAction(object):
         return {
             'action_type': self.action_type,
             'action_customization_args': (
-                schema_utils.CustomizationArgsUtil.get_full_customization_args(
+                customization_args_util.get_full_customization_args(
                     self.action_customization_args,
                     action_registry.Registry.get_action_by_type(
                         self.action_type).customization_arg_specs)),
@@ -869,7 +869,7 @@ class LearnerAction(object):
             raise utils.ValidationError(
                 'Invalid action type: %s' % self.action_type)
 
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
+        customization_args_util.validate_customization_args_and_values(
             'action', self.action_type, self.action_customization_args,
             action.customization_arg_specs)
