+++ schema_utils.py
@@ -24,14 +24,12 @@ The objects that can be described by these schemas must be composable from the
 following Python types: bool, dict, float, int, list, unicode.
 """
 
-import logging
 import numbers
 import re
 import urllib
 import urlparse
 
 from core.domain import html_cleaner  # pylint: disable=relative-import
-import utils # pylint: disable=relative-import
 
 SCHEMA_KEY_ITEMS = 'items'
 SCHEMA_KEY_LEN = 'len'
@@ -375,108 +373,3 @@ class _Validators(object):
             bool. Whether the given object is a valid email.
         """
         return bool(re.search(r'^[\w\.\+\-]+\@[\w]+\.[a-z]{2,3}$', obj))
-
-
-class CustomizationArgsUtil(object):
-    """Utility class for obtaining and validating customization args
-    for interactions.
-    """
-
-    @staticmethod
-    def get_full_customization_args(customization_args, ca_specs):
-        """Populates the given customization_args dict with default values
-        if any of the expected customization_args are missing.
-
-        Args:
-            customization_args: dict. The customization dict. The keys are names
-                of customization_args and the values are dicts with a
-                single key, 'value', whose corresponding value is the value of
-                the customization arg.
-            ca_specs: list(dict). List of spec dictionaries. Is used to check if
-                some keys are missing in customization_args. Dicts have the
-                following structure:
-                    - name: str. The customization variable name.
-                    - description: str. The customization variable description.
-                    - default_value: *. The default value of the customization
-                        variable.
-
-        Returns:
-            dict. The customization_args dict where missing keys are populated
-            with the default values.
-        """
-        for ca_spec in ca_specs:
-            if ca_spec.name not in customization_args:
-                customization_args[ca_spec.name] = {
-                    'value': ca_spec.default_value
-                }
-        return customization_args
-
-    @staticmethod
-    def validate_customization_args_and_values(
-            item_name, item_type, customization_args,
-            ca_specs_to_validate_against):
-        """Validates the given `customization_args` dict against the specs set
-        out in 'ca_specs_to_validate_against'. 'item_name' and 'item_type' are
-        used to populate any error messages that arise during validation.
-        Note that this may modify the given customization_args dict, if it has
-        extra or missing keys. It also normalizes any HTML in the
-        customization_args dict.
-
-        Args:
-            item_name: str. This is always 'interaction'.
-            item_type: str. The item_type is the ID of the interaction.
-            customization_args: dict. The customization dict. The keys are names
-                of customization_args and the values are dicts with a
-                single key, 'value', whose corresponding value is the value of
-                the customization arg.
-            ca_specs_to_validate_against: list(dict). List of spec dictionaries.
-                Is used to check if some keys are missing in customization_args.
-                Dicts have the following structure:
-                    - name: str. The customization variable name.
-                    - description: str. The customization variable description.
-                    - default_value: *. The default value of the customization
-                        variable.
-
-        Raises:
-            ValidationError: The given 'customization_args' is not valid.
-        """
-        ca_spec_names = [
-            ca_spec.name for ca_spec in ca_specs_to_validate_against]
-
-        if not isinstance(customization_args, dict):
-            raise utils.ValidationError(
-                'Expected customization args to be a dict, received %s'
-                % customization_args)
-
-        # Validate and clean up the customization args.
-
-        # Populate missing keys with the default values.
-        customization_args = CustomizationArgsUtil.get_full_customization_args(
-            customization_args, ca_specs_to_validate_against)
-
-        # Remove extra keys.
-        extra_args = []
-        for arg_name in customization_args.keys():
-            if not isinstance(arg_name, basestring):
-                raise utils.ValidationError(
-                    'Invalid customization arg name: %s' % arg_name)
-            if arg_name not in ca_spec_names:
-                extra_args.append(arg_name)
-                logging.warning(
-                    '%s %s does not support customization arg %s.'
-                    % (item_name.capitalize(), item_type, arg_name))
-        for extra_arg in extra_args:
-            del customization_args[extra_arg]
-
-        # Check that each value has the correct type.
-        for ca_spec in ca_specs_to_validate_against:
-            try:
-                customization_args[ca_spec.name]['value'] = (
-                    normalize_against_schema(
-                        customization_args[ca_spec.name]['value'],
-                        ca_spec.schema))
-            except Exception:
-                # TODO(sll): Raise an actual exception here if parameters are
-                # not involved (If they are, can we get sample values for the
-                # state context parameters?).
-                pass
