+++ schema_utils_test.py
@@ -19,13 +19,10 @@
 # pylint: disable=relative-import
 
 import inspect
-import logging
 
 from core.domain import email_manager
-from core.domain import interaction_registry
 from core.tests import test_utils
 import schema_utils
-import utils
 
 SCHEMA_KEY_ITEMS = schema_utils.SCHEMA_KEY_ITEMS
 SCHEMA_KEY_LEN = schema_utils.SCHEMA_KEY_LEN
@@ -584,313 +581,3 @@ class SchemaNormalizationUnitTests(test_utils.GenericTestBase):
         invalid_vals = [[u'admin@oppia'], big_email_list,
                         [u'admin@oppia.commmm'], [u'a@.com']]
         self.check_normalization(schema, mappings, invalid_vals)
-
-
-class CustomizationArgsUtilUnitTests(test_utils.GenericTestBase):
-    """Test customization args generation and validation."""
-
-    def test_get_full_customization_args(self):
-        """Test get full customization args method."""
-        ca_continue_specs = interaction_registry.Registry.get_interaction_by_id(
-            'Continue').customization_arg_specs
-
-        complete_customization_args = {
-            'buttonText': {'value': 'Please Continue'}
-        }
-
-        complete_customization_args_with_extra_arg = {
-            'buttonText': {'value': 'Please Continue'},
-            'extraArg': {'value': ''}
-        }
-
-        # Check if no new key is added to customization arg dict if all specs
-        # are present.
-        self.assertEqual(
-            complete_customization_args,
-            schema_utils.CustomizationArgsUtil.get_full_customization_args(
-                complete_customization_args, ca_continue_specs
-            )
-        )
-
-        # Check if no new key is added to customization arg dict and extra keys
-        # are not removed if all specs are present.
-        self.assertEqual(
-            complete_customization_args_with_extra_arg,
-            schema_utils.CustomizationArgsUtil.get_full_customization_args(
-                complete_customization_args_with_extra_arg,
-                ca_continue_specs
-            )
-        )
-
-        ca_fraction_input_specs = (
-            interaction_registry.Registry.get_interaction_by_id(
-                'FractionInput').customization_arg_specs
-        )
-
-        incomplete_customization_args = {
-            'requireSimplestForm': {'value': False},
-            'allowNonzeroIntegerPart': {'value': False}
-        }
-
-        incomplete_customization_args_with_extra_arg = {
-            'requireSimplestForm': {'value': False},
-            'allowNonzeroIntegerPart': {'value': False},
-            'extraArg': {'value': ''}
-        }
-
-        expected_complete_customization_args = {
-            'requireSimplestForm': {'value': False},
-            'allowImproperFraction': {'value': True},
-            'allowNonzeroIntegerPart': {'value': False},
-            'customPlaceholder': {'value': ''}
-        }
-
-        expected_complete_customization_args_with_extra_arg = {
-            'requireSimplestForm': {'value': False},
-            'allowImproperFraction': {'value': True},
-            'allowNonzeroIntegerPart': {'value': False},
-            'customPlaceholder': {'value': ''},
-            'extraArg': {'value': ''}
-        }
-
-        # Check if missing specs are added to customization arg dict without
-        # making any other change.
-        self.assertEqual(
-            expected_complete_customization_args,
-            schema_utils.CustomizationArgsUtil.get_full_customization_args(
-                incomplete_customization_args, ca_fraction_input_specs
-            )
-        )
-
-        # Check if missing specs are added to customization arg dict without
-        # making any other change and without removing extra args.
-        self.assertEqual(
-            expected_complete_customization_args_with_extra_arg,
-            schema_utils.CustomizationArgsUtil.get_full_customization_args(
-                incomplete_customization_args_with_extra_arg,
-                ca_fraction_input_specs
-            )
-        )
-
-    def test_validate_customization_args_and_values(self):
-        """Test validate customization args and values method."""
-
-        observed_log_messages = []
-
-        def mock_logging_function(msg, *_):
-            observed_log_messages.append(msg)
-
-        ca_item_selection_specs = (
-            interaction_registry.Registry.get_interaction_by_id(
-                'ItemSelectionInput').customization_arg_specs
-        )
-
-        complete_customization_args = {
-            'minAllowableSelectionCount': {'value': 1},
-            'maxAllowableSelectionCount': {'value': 1},
-            'choices': ['']
-        }
-
-        complete_customization_args_with_invalid_arg_name = {
-            'minAllowableSelectionCount': {'value': 1},
-            'maxAllowableSelectionCount': {'value': 1},
-            'choices': [''],
-            23: {'value': ''}
-        }
-
-        complete_customization_args_with_extra_arg = {
-            'minAllowableSelectionCount': {'value': 1},
-            'maxAllowableSelectionCount': {'value': 1},
-            'choices': [''],
-            'extraArg': {'value': ''}
-        }
-
-        complete_customization_args_with_invalid_arg_type = {
-            'minAllowableSelectionCount': {'value': 'invalid'},
-            'maxAllowableSelectionCount': {'value': 1},
-            'choices': ['']
-        }
-
-        expected_customization_args_after_validation = {
-            'minAllowableSelectionCount': {'value': 1},
-            'maxAllowableSelectionCount': {'value': 1},
-            'choices': ['']
-        }
-
-        expected_customization_args_after_validation_with_invalid_arg_type = (
-            complete_customization_args_with_invalid_arg_type
-        )
-
-        # The next four checks are for cases where customization args dict
-        # contains all required specs.
-
-        # Check if no error is produced for valid customization args.
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-            'interaction',
-            'ItemSelectionInput',
-            complete_customization_args,
-            ca_item_selection_specs
-        )
-        self.assertEqual(
-            expected_customization_args_after_validation,
-            complete_customization_args
-        )
-
-        # Check if error is produced when arg name is invalid.
-        with self.assertRaisesRegexp(
-            utils.ValidationError,
-            'Invalid customization arg name: 23'
-        ):
-            schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-                'interaction',
-                'ItemSelectionInput',
-                complete_customization_args_with_invalid_arg_name,
-                ca_item_selection_specs
-            )
-
-        # Check if error is produced when extra args are present.
-        with self.swap(logging, 'warning', mock_logging_function):
-            schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-                'interaction',
-                'ItemSelectionInput',
-                complete_customization_args_with_extra_arg,
-                ca_item_selection_specs
-            )
-            self.assertEqual(len(observed_log_messages), 1)
-            self.assertEqual(
-                observed_log_messages[0],
-                (
-                    'Interaction ItemSelectionInput does not support '
-                    'customization arg extraArg.'
-                )
-            )
-            self.assertEqual(
-                expected_customization_args_after_validation,
-                complete_customization_args_with_extra_arg
-            )
-
-        # Check if no error is produced when arg type is not valid.
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-            'interaction',
-            'ItemSelectionInput',
-            complete_customization_args_with_invalid_arg_type,
-            ca_item_selection_specs
-        )
-
-        self.assertEqual(
-            expected_customization_args_after_validation_with_invalid_arg_type,
-            complete_customization_args_with_invalid_arg_type
-        )
-
-        ca_fraction_input_specs = (
-            interaction_registry.Registry.get_interaction_by_id(
-                'FractionInput').customization_arg_specs
-        )
-
-        incomplete_customization_args = {
-            'requireSimplestForm': {'value': False},
-            'allowNonzeroIntegerPart': {'value': False}
-        }
-
-        incomplete_customization_args_with_invalid_arg_name = {
-            'requireSimplestForm': {'value': False},
-            False: {'value': False},
-        }
-
-        incomplete_customization_args_with_extra_arg = {
-            'requireSimplestForm': {'value': False},
-            'allowNonzeroIntegerPart': {'value': False},
-            'extraArg': {'value': ''}
-        }
-
-        incomplete_customization_args_with_invalid_arg_type = {
-            'requireSimplestForm': {'value': False},
-            'allowNonzeroIntegerPart': {'value': False},
-            'customPlaceholder': {'value': 12}
-        }
-
-        expected_customization_args_after_validation = {
-            'requireSimplestForm': {'value': False},
-            'allowImproperFraction': {'value': True},
-            'allowNonzeroIntegerPart': {'value': False},
-            'customPlaceholder': {'value': ''}
-        }
-
-        expected_customization_args_after_validation_with_invalid_arg_type = (
-            incomplete_customization_args_with_invalid_arg_type
-        )
-
-        # The next four checks are for cases where customization args dict
-        # does not contain some of the required specs.
-
-        # Check if no error is produced for valid customization args.
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-            'interaction',
-            'FractionInput',
-            incomplete_customization_args,
-            ca_fraction_input_specs
-        )
-        self.assertEqual(
-            expected_customization_args_after_validation,
-            incomplete_customization_args
-        )
-
-        # Check if error is produced when arg name is invalid.
-        with self.assertRaisesRegexp(
-            utils.ValidationError,
-            'Invalid customization arg name: False'
-        ):
-            schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-                'interaction',
-                'FractionInput',
-                incomplete_customization_args_with_invalid_arg_name,
-                ca_fraction_input_specs
-            )
-
-        # Check if error is produced when extra args are present.
-        with self.swap(logging, 'warning', mock_logging_function):
-            schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-                'interaction',
-                'FractionInput',
-                incomplete_customization_args_with_extra_arg,
-                ca_fraction_input_specs
-            )
-            self.assertEqual(len(observed_log_messages), 2)
-            self.assertEqual(
-                observed_log_messages[1],
-                (
-                    'Interaction FractionInput does not support customization '
-                    'arg extraArg.'
-                )
-            )
-            self.assertEqual(
-                expected_customization_args_after_validation,
-                incomplete_customization_args_with_extra_arg
-            )
-
-        # Check if no error is produced when arg type is not valid.
-        schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-            'interaction',
-            'FractionInput',
-            incomplete_customization_args_with_invalid_arg_type,
-            ca_fraction_input_specs
-        )
-        self.assertEqual(
-            expected_customization_args_after_validation_with_invalid_arg_type,
-            incomplete_customization_args_with_invalid_arg_type
-        )
-
-        # A general check to see if error are produced when customization args
-        # is not of type dict.
-        customization_args_with_invalid_type = 23
-        with self.assertRaisesRegexp(
-            utils.ValidationError,
-            'Expected customization args to be a dict, received %s'
-            % customization_args_with_invalid_type
-        ):
-            schema_utils.CustomizationArgsUtil.validate_customization_args_and_values( #pylint: disable=line-too-long
-                'interaction',
-                'FractionInput',
-                customization_args_with_invalid_type,
-                ca_fraction_input_specs
-            )
